zabbia/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.8
pydantic==2.4.2
aiohttp==3.8.6
grpcio==1.59.2
grpcio-tools==1.59.2
python-jose==3.3.0
python-multipart==0.0.6
asyncpg==0.28.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.0.335
langchain-community==0.0.13
pyjwt==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1 